+++ nitrocli/src/tests/status.rs
@@ -37,14 +37,40 @@ fn not_found() {
 }
 
 #[test_device]
-fn output(device: nitrokey::DeviceWrapper) -> crate::Result<()> {
+fn output_pro(device: nitrokey::Pro) -> crate::Result<()> {
   let re = regex::Regex::new(
     r#"^Status:
-  model:             (Pro|Storage)
+  model:             Pro
   serial number:     0x[[:xdigit:]]{8}
   firmware version:  \d+\.\d+
   user retry count:  [0-3]
   admin retry count: [0-3]
+$"#,
+  )
+  .unwrap();
+
+  let out = Nitrocli::with_dev(device).handle(&["status"])?;
+  assert!(re.is_match(&out), out);
+  Ok(())
+}
+
+#[test_device]
+fn output_storage(device: nitrokey::Storage) -> crate::Result<()> {
+  let re = regex::Regex::new(
+    r#"^Status:
+  model:             Storage
+  serial number:     0x[[:xdigit:]]{8}
+  firmware version:  \d+\.\d+
+  user retry count:  [0-3]
+  admin retry count: [0-3]
+  Storage:
+    SD card ID:        0x[[:xdigit:]]{8}
+    firmware:          (un)?locked
+    storage keys:      (not )?created
+    volumes:
+      unencrypted:     (read-only|active|inactive)
+      encrypted:       (read-only|active|inactive)
+      hidden:          (read-only|active|inactive)
 $"#,
   )
   .unwrap();