+++ nitrocli/CHANGELOG.md
@@ -1,5 +1,11 @@
 Unreleased
 ----------
+- Removed `storage status` subcommand
+  - Moved its output into `status` command
+
+
+Unreleased
+----------
 - Added the `reset` command to perform a factory reset
 - Added the `-V`/`--version` option to print the program's version
 - Check the status of a PWS slot before accessing it in `pws get`